It Runs!
So still a lot of work to do but it runs...the mainwindow loads and you
have the "file" menu and the connection manager etc etc the settings all
load from config bla bla but deffinataly several errors seen in cmd
window and as you can see i just commented out the tray icon code. But
hey at least it starts up and is some what functional.