[Torrent] Add trackers cache
The current way the trackers are being update is either by adding a new
torrent to the session (on startup or new one), or by changing the
trackers.
This leads to data not being updated in the `torrent.trackers` dict.
To solve this, a cache mechanism was added which will update the dict
on changes or every 5 seconds.
This will help us also get an update regarding `lt.announce_endpoint` in
 each of the trackers.